Adiciona tratamento basico de exceções e modifica um pouco a documentação

+++ ApiCatalogoPersonagens/Services/PersonagemService.cs
@@ -25,38 +25,29 @@ namespace ApiCatalogoPersonagens.Services
 
         public async Task AtualizarPersonagem(Guid idPersonagem, PersonagemInputModel personagemInputModel)
         {
-            // Procura no banco um personagem com o id selecionado
+
             var personagemEntity = await _personagemRepository.ObterPersonagem(idPersonagem);
 
-            // Se o personagem não existir lança excessão
             if (personagemEntity == null) throw new PersonagemNaoCadastradoException();
 
-            // Atualiza os dados do personagem selecionado
             personagemEntity = _mapper.Map<Personagem>(personagemInputModel);
             personagemEntity.Id = idPersonagem;
-            // Atualiza os dados no banco de dados
+
             await _personagemRepository.AtualizarPersonagem(personagemEntity);
         }
 
         public async Task<PersonagemViewModel> InserirPersonagem(PersonagemInputModel personagemInputModel)
         {
-            // Verifica se o personagem ja existe no banco de dados
-          //  var personagemEntity = await _personagemRepository.ObterPersonagem(personagemInputModel.Nome, personagemInputModel.Filme);
-            
-            // Se o personagem existir é lancada uma exceção
-            //if (personagemEntity != null)
-           // {
-             //   throw new PersonagemJaCadastradoException();
-            //}
+           
+             var personagemEntity = await _personagemRepository.ObterPersonagem(personagemInputModel.Nome, personagemInputModel.Filme);
 
-            // Mapeia os dados da entidade para um modelo de input
+            if (personagemEntity != null) throw new PersonagemJaCadastradoException();
+
             var personagemInsert = _mapper.Map<Personagem>(personagemInputModel);
             personagemInsert.Id = Guid.NewGuid();
             
-            // Faz a inserção dos dados no Banco
             await _personagemRepository.InserirPersonagem(personagemInsert);
             
-            // Retorna um modelo que o usuário pode ter acesso
             return _mapper.Map<PersonagemViewModel>(personagemInsert);
         }
 
@@ -64,7 +55,7 @@ namespace ApiCatalogoPersonagens.Services
         {
             var personagem = await _personagemRepository.ObterPersonagem(idPersonagem);
 
-            if (personagem == null) return null;
+            if (personagem == null) throw new PersonagemNaoCadastradoException();
 
             else return _mapper.Map<PersonagemViewModel>(personagem);
         }
@@ -73,7 +64,7 @@ namespace ApiCatalogoPersonagens.Services
         {
             var personagem = await _personagemRepository.ObterPersonagem(nome,filme);
 
-            if (personagem == null) return null;
+            if (personagem == null) throw new PersonagemNaoCadastradoException();
 
             else return _mapper.Map<PersonagemViewModel>(personagem);
         }
@@ -81,22 +72,29 @@ namespace ApiCatalogoPersonagens.Services
         public async Task<List<PersonagemViewModel>> ObterPFilme(int pagina, int quantidade, string filme)
         {
             var personagens = await _personagemRepository.ObterPFilme(pagina, quantidade, filme);
+            if (personagens.Count == 0) throw new PersonagemNaoCadastradoException();
             return _mapper.Map<List<PersonagemViewModel>>(personagens);
         }
 
         public async Task<List<PersonagemViewModel>> ObterPNome(int pagina, int quantidade, string nome)
         {
             var personagens = await _personagemRepository.ObterPNome(pagina, quantidade, nome);
+           
+            if (personagens == null) throw new PersonagemNaoCadastradoException();
+            
             return _mapper.Map<List<PersonagemViewModel>>(personagens);
         }
 
         public async Task<List<PersonagemViewModel>> ObterTodos(int pagina, int quantidade)
         {
+            
             var personagens = await _personagemRepository.ObterTodos(pagina, quantidade);
+            
+            if (personagens.Count == 0) throw new NenhumPersonagemCadastradoException();
+            
             return _mapper.Map<List<PersonagemViewModel>>(personagens);
         }
 
-        //Fecha a conexão com o banco de dados
         public void Dispose()
         {
             _personagemRepository?.Dispose();
@@ -104,6 +102,10 @@ namespace ApiCatalogoPersonagens.Services
 
         public async Task RemoverPersonagem(Guid idPersonagem)
         {
+            var personagemEntity = await _personagemRepository.ObterPersonagem(idPersonagem);
+
+            if (personagemEntity == null) throw new PersonagemNaoCadastradoException();
+
             await _personagemRepository.RemoverPersonagem(idPersonagem); 
         }
     }
